deepfake_monitor_repo/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
jinja2==3.1.6
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
celery==5.5.3
redis==6.4.0
python-multipart==0.0.20
opencv-python-headless==4.12.0.88
numpy==2.2.6
scikit-learn==1.7.1
ffmpeg-python==0.2.0
openai-whisper==20250625
insightface==0.7.3
onnxruntime==1.22.1
pytesseract==0.3.13
Pillow==11.3.0
python-dotenv==1.1.1
python-slugify==8.0.4
